Add an extra feature and remove unneccessary commands
This update adds a feature to check if a repo in the list is shallow and unshallow it from the origin of it.
Also the method to check if the repo contains AOSP remote and the URL of it is valid has been finetuned.
Now it doesn't try to fetch the tags at all costs, only after the script made sure the already existing AOSP remote
is pointing to a valid URL