Drag&Drop>Drag&F@ck
Drag&Drop adds a ' to the beginning of the path and a ' to the end of it.
Previously this was handled by a graceful f*cking crash...ˇ_ˇ
Now it just cuts the first and last characters off of the input
string if it contains \' or ' making it valid to the "isPath" check
later and and accepts the drag&dropped path as valid source root.